Refactored Home.tsx, broke out subcomponents
