+++ frontend-react/src/pages/Home/Interfaces.ts
@@ -1,0 +1,20 @@
+export interface SectionProp {
+    title?: string;
+    type?: string;
+    summary?: string;
+    bullets?: { content?: string }[];
+    features?: FeatureProp[];
+    description?: string;
+    buttonText?: string;
+    buttonUrlSubject?: string;
+}
+
+export interface FeatureProp {
+    method?: number;
+    title?: string;
+    icon?: string;
+    img?: string;
+    imgAlt?: string;
+    summary?: string;
+    items?: { title?: string; summary?: string }[];
+}
